all: implement cell formatting
One of the most hardest task to finished after inline markup. It take
three refactoring until we get the current implementation.

To make the cell formatting works, we breaking down all table contents
into cells.  When parsing a cells, we lookup the content before '|' and
after '\n'. If its a valid cell format, it will be set as cell format
for the next cell, otherwise it will considered as normal text.