all: fix parsing and rendering cross reference
Previously, when parsing cross reference we assume that if the string
contains upper-case letter then it's a label so we store it as title
to search for ID later.

The bug is when ID is set manually and its contains upper-case for
example "[#Id]".

This changes fix this issue by storing cross reference first, not
assuming it as ID or title, and then when doing rendering we check
whether its ID or title.