all: refactoring handling generate ref ID
Previously, we always set the ID prefix and separator default to "_" if
its not set, this cause all of the ID is prefixed with "_".

This changes use strict rules when generating ID following the Mozilla
specification [1] and latest AsciiDoc Language [2].

The idprefix must be ASCII string.
It must start with "_", "-", or ASCII letters, otherwise the "_" will
be added to the beginning.
If one of the character is not valid, it will replaced with "_".

The `idseparator` can be empty or single ASCII character ('_' or '-',
ASCII letter, or digit).
It is used to replace invalid characters in the REF_ID.

[1] https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/id
[2] https://docs.asciidoctor.org/asciidoc/latest/sections/id-prefix-and-separator/