all: add support for unordered list with '-'
The unordered list item with hyphen ('-') cause too much confusion and
inconsistency, nevertheless most of still use it.

Case one, given the following markup,

```
- Item 1
+
"A line
of quote"
-- Author
```

Is the "Author" the sub item in list or we are parsing author of quote
paragraph?

Case two, the writer want to write em dash (`&#8212;` in HTML Unicode) but
somehow the editor wrap it and start in new line.

As a reminder, the official documentation only recommend using hyphen for
simple list item [1].

[1] https://docs.asciidoctor.org/asciidoc/latest/lists/unordered/#basic-unordered-list