all: replace linter golangci-lint with revive and shadow
The golangci-lint does not output any results anymore
Either we are getting good at writing Go or the linter itself is become
less good.
We also have seen that the latest golangci-lint is failed to build
with Go tip, a simple "make" on the golangci-lint never success in my
experiences.

This changes fix all the output reported by the revive and shadow.