Release asciidoctor-go v0.7.0 (2025-04-18)
[NEW FEATURE] Support include directive inside block code.

Example of block code with include directive,

----
...
include::file[]
...
----

[NEW FEATURE] Support document attribute "docdir"

The "docdir" attribute contains the full path of the directory that
contains the source document.
By default it is set to the directory where the Document resided.

[NEW FEATURE] Add default HTML stylesheet.

The generated HTML now contains the default stylesheet.
The stylesheet is copied from HTML file generated by Asciidoctor
v2.0.23.